feat(85-05): wire EndgameOverallPerformanceSection in Endgames.tsx, rename test files
- Endgames.tsx: swap EndgameGamesWithWithoutSection + EndgameStartVsEndSection
  imports/mounts for single EndgameOverallPerformanceSection mount
- Create EndgameOverallPerformanceSection.test.tsx: 3-card layout, Card 2 rows,
  score-gap, sig-gating, empty-state, and WDL-bar-absence assertions
- Rename Endgames.startVsEnd.test.tsx -> Endgames.overallPerformance.test.tsx:
  update for single composite section mount, 3-card testids, accordion-order tests
- Delete EndgameGamesWithWithoutSection.test.tsx and EndgameStartVsEndSection.test.tsx
- Update AchievableScorePopover.tsx comment to reference new component name
- All 5 frontend gates pass: lint, tsc, vitest (343), knip, build
